# docs/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Speedtest Watcher API",
    "description": "Results, statistics, settings, backups and metrics of a Speedtest Watcher instance.\n\n- **Authentication:** while sign-in is off, no token is needed. While it's on, create a token under **Settings → Security → API Token** and send it as `Authorization: Bearer swt_…`. Read operations also work without a token when people who aren't signed in have read-only access.\n- **Errors** answer with `{ \"message\": \"…\" }` and a 4xx status. `401` means the request needs a valid token.\n- **Units:** timestamps are UTC ISO 8601, speeds are Mbps, ping and jitter are milliseconds, and test durations are seconds. Failed and skipped results have `-1` readings.\n\nThe guide, with examples, is at https://github.com/remotesojourner/speedtest-watcher/blob/main/docs/api.md.",
    "license": {
      "name": "AGPL-3.0-only",
      "identifier": "AGPL-3.0-only"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/speedtests": {
      "get": {
        "tags": [
          "Speedtests"
        ],
        "summary": "List results",
        "description": "Returns results newest first. To get the next page, pass the `id` of the last result you received as `afterId`. An `afterId` that doesn't exist gives an empty list.\n\n**Access:** read. Works for everyone while sign-in is off, and for read-only visitors, signed-in users and API tokens while it's on.",
        "parameters": [
          {
            "name": "afterId",
            "in": "query",
            "description": "Continue after this result.",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "How many results to return. 0 or less gives 10.",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Only results with this status.",
            "schema": {
              "$ref": "#/components/schemas/TestStatus"
            }
          },
          {
            "name": "type",
            "in": "query",
            "description": "Only results started this way: `auto` by the schedule, `custom` by hand.",
            "schema": {
              "$ref": "#/components/schemas/TestType"
            }
          },
          {
            "name": "healthy",
            "in": "query",
            "description": "Only results that met (`true`) or missed (`false`) their targets.",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The results.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SpeedtestResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "A filter has a value that isn't allowed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Sign-in is on, people who aren't signed in have no access, and the request has no valid API token or signed-in session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-access": "read"
      }
    },
    "/api/speedtests/{id}": {
      "get": {
        "tags": [
          "Speedtests"
        ],
        "summary": "Get a result",
        "description": "**Access:** read. Works for everyone while sign-in is off, and for read-only visitors, signed-in users and API tokens while it's on.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The result's id.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The result.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpeedtestResponse"
                }
              }
            }
          },
          "404": {
            "description": "No result has this id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Sign-in is on, people who aren't signed in have no access, and the request has no valid API token or signed-in session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-access": "read"
      },
      "delete": {
        "tags": [
          "Speedtests"
        ],
        "summary": "Delete a result",
        "description": "**Access:** full. Needs sign-in to be off, a signed-in browser session, or an API token.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The result's id.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The result was deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "404": {
            "description": "No result has this id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Sign-in is on and the request has no valid API token or signed-in session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-access": "full"
      }
    },
    "/api/speedtests/statistics": {
      "get": {
        "tags": [
          "Speedtests"
        ],
        "summary": "Get statistics",
        "description": "Summarises the tests in a period: counts, the lowest, average and highest readings, how steady they were, averages per hour of the day, and points for charts.\n\n`from` and `to` each take a date (`2026-09-14`), which covers that whole day, or a date and time. A time without an offset is read in `tz`. A value that can't be read falls back to the default.\n\n**Access:** read. Works for everyone while sign-in is off, and for read-only visitors, signed-in users and API tokens while it's on.",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "description": "The start of the period. Defaults to seven days ago.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "in": "query",
            "description": "The end of the period. Defaults to today.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tz",
            "in": "query",
            "description": "An IANA time zone, such as `Europe/London`, for day boundaries and the hourly averages. Defaults to UTC.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The statistics.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatisticsResponse"
                }
              }
            }
          },
          "400": {
            "description": "The time zone isn't one the server knows.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Sign-in is on, people who aren't signed in have no access, and the request has no valid API token or signed-in session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-access": "read"
      }
    },
    "/api/speedtests/export": {
      "post": {
        "tags": [
          "Speedtests"
        ],
        "summary": "Export results",
        "description": "Downloads the matching results as a CSV or JSON file, oldest first. A JSON export can be imported again through the Storage tab's Import data, or wrapped as `{ \"speedtests\": [...] }` and sent to `PUT /api/storage/data`.\n\n**Access:** read. Works for everyone while sign-in is off, and for read-only visitors, signed-in users and API tokens while it's on.",
        "requestBody": {
          "description": "Which results to export and in what format.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportResultsRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ExportResultsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The file.",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "The request body isn't valid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Sign-in is on, people who aren't signed in have no access, and the request has no valid API token or signed-in session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-access": "read"
      }
    },
    "/api/speedtests/status": {
      "get": {
        "tags": [
          "Speedtests"
        ],
        "summary": "Get the run status",
        "description": "**Access:** read. Works for everyone while sign-in is off, and for read-only visitors, signed-in users and API tokens while it's on.",
        "responses": {
          "200": {
            "description": "Whether a test is running and whether tests are paused.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunStatusResponse"
                }
              }
            }
          },
          "401": {
            "description": "Sign-in is on, people who aren't signed in have no access, and the request has no valid API token or signed-in session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-access": "read"
      }
    },
    "/api/speedtests/run": {
      "post": {
        "tags": [
          "Speedtests"
        ],
        "summary": "Start a speedtest",
        "description": "Starts a manual test and answers straight away, while the test runs in the background. Poll `GET /api/speedtests/status` or the result list to see when it has finished.\n\n**Access:** full. Needs sign-in to be off, a signed-in browser session, or an API token.",
        "parameters": [
          {
            "name": "serverId",
            "in": "query",
            "description": "Test against this server instead of the configured choice. Only providers with server choice use it.",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The test has started.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "400": {
            "description": "The server id isn't a number.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "A test is already running, tests are paused, or no provider is chosen.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Sign-in is on and the request has no valid API token or signed-in session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-access": "full"
      }
    },
    "/api/speedtests/pause": {
      "post": {
        "tags": [
          "Speedtests"
        ],
        "summary": "Pause speedtests",
        "description": "Stops scheduled and manual tests. A test that is already running finishes. The pause is kept in memory, so restarting the app resumes tests.\n\n**Access:** full. Needs sign-in to be off, a signed-in browser session, or an API token.",
        "requestBody": {
          "description": "How long to pause. Send no body, or leave `resumeIn` out, to pause until tests are resumed.",
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/PauseRequest"
                  }
                ]
              }
            },
            "application/*+json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/PauseRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tests are paused.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "400": {
            "description": "The pause is longer than 720 hours.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Sign-in is on and the request has no valid API token or signed-in session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-access": "full"
      }
    },
    "/api/speedtests/continue": {
      "post": {
        "tags": [
          "Speedtests"
        ],
        "summary": "Resume speedtests",
        "description": "**Access:** full. Needs sign-in to be off, a signed-in browser session, or an API token.",
        "responses": {
          "200": {
            "description": "Tests are running on schedule again.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "401": {
            "description": "Sign-in is on and the request has no valid API token or signed-in session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-access": "full"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "required": [
          "message"
        ],
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "What went wrong, written for people."
          }
        },
        "description": "Why a request failed. Every error the API answers carries this body."
      },
      "ExportResultsRequest": {
        "type": "object",
        "properties": {
          "format": {
            "examples": [
              "json"
            ],
            "type": "string",
            "description": "`csv` or `json`. Anything else gives CSV. The JSON can be imported again through the Storage tab's Import data."
          },
          "status": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TestStatus"
              }
            ],
            "description": "Only results with this status."
          },
          "type": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TestType"
              }
            ],
            "description": "Only results started this way."
          },
          "healthy": {
            "type": [
              "null",
              "boolean"
            ],
            "description": "Only results that met (`true`) or missed (`false`) their targets."
          },
          "ids": {
            "examples": [
              [
                41,
                42
              ]
            ],
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "description": "Only these results. The other filters still apply."
          }
        },
        "description": "Which results to export and in what format. Leave everything out to export every result as CSV."
      },
...
      "MessageResponse": {
        "required": [
          "message"
        ],
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "What happened, written for people."
          }
        },
        "description": "Confirms that a request which changes something succeeded."
      },
...
      "StatisticsResponse": {
        "required": [
          "dateRange",
          "tests",
          "ping",
          "jitter",
          "download",
          "upload",
          "time",
          "consistency",
          "packetLoss",
          "bufferbloatDown",
          "bufferbloatUp",
          "dataUsedBytes",
          "hourlyAverages",
          "points",
          "rawDataPoints",
          "downsampled"
        ],
        "type": "object",
        "properties": {
          "dateRange": {
            "description": "The period the figures cover.",
            "$ref": "#/components/schemas/StatisticsPeriodResponse"
          },
          "tests": {
            "description": "How many tests ran in the period and how many of them failed.",
            "$ref": "#/components/schemas/TestCountsResponse"
          },
          "ping": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/MetricSummaryResponse"
              }
            ],
            "description": "Ping of the completed tests, in milliseconds. `null` when none completed."
          },
          "jitter": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/MetricSummaryResponse"
              }
            ],
            "description": "Jitter of the completed tests, in milliseconds. `null` when none reported jitter."
          },
          "download": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/MetricSummaryResponse"
              }
            ],
            "description": "Download speed of the completed tests, in Mbps. `null` when none completed."
          },
          "upload": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/MetricSummaryResponse"
              }
            ],
            "description": "Upload speed of the completed tests, in Mbps. `null` when none completed."
          },
          "time": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/MetricSummaryResponse"
              }
            ],
            "description": "How long the completed tests took, in seconds. `null` when none completed."
          },
          "consistency": {
            "description": "How steady the completed tests' readings were.",
            "$ref": "#/components/schemas/ConsistencyResponse"
          },
          "packetLoss": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/MetricSummaryResponse"
              }
            ],
            "description": "Packet loss across the completed tests that measured it, as percentages. Only Ookla measures it."
          },
          "bufferbloatDown": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/MetricSummaryResponse"
              }
            ],
            "description": "Bufferbloat while downloading, in milliseconds, across the completed tests that have a figure, or `null` when none do."
          },
          "bufferbloatUp": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/MetricSummaryResponse"
              }
            ],
            "description": "Bufferbloat while uploading, in milliseconds, across the completed tests that have a figure, or `null` when none do."
          },
          "dataUsedBytes": {
            "examples": [
              991777425
            ],
            "type": "integer",
            "description": "Bytes the tests in the period moved, download and upload together, as far as the providers reported them.",
            "format": "int64"
          },
          "hourlyAverages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HourlyAverageResponse"
            },
            "description": "Averages of the completed tests for each hour of the day, 0 to 23, in the requested time zone."
          },
          "points": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChartPointResponse"
            },
            "description": "Chart points, oldest first. Up to 300 tests give one point each. More are averaged into 300 equal slices of the period, and slices without tests are left out."
          },
          "rawDataPoints": {
            "examples": [
              4
            ],
            "type": "integer",
            "description": "How many tests ran in the period, before any averaging.",
            "format": "int32"
          },
          "downsampled": {
            "type": "boolean",
            "description": "Whether the points were averaged down from more tests."
          }
        },
        "description": "Summary figures and chart points for a period."
      },
...
      "TestStatus": {
        "enum": [
          "completed",
          "failed",
          "skipped"
        ]
      },
      "TestType": {
        "enum": [
          "auto",
          "custom"
        ]
      },
...
    },
    "securitySchemes": {
      "apiToken": {
        "type": "http",
        "description": "An API token (`swt_…`) from **Settings → Security → API Token**. It's needed only while sign-in is on, and read operations don't need it when people who aren't signed in have read-only access.",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "apiToken": [ ]
    },
    { }
  ],
...
}